Support loading monochrome images as gray scale
This reduces memory usage by 2/3 for images without an alpha channel
and 1/2 for images with an alpha channel.

The UnpremultiplyColor overload that returns a float will no longer
round the value to the nearest integer before returning it.
The rounding was causing incorrect behavior when processing floating
point image data in the [0, 1] range.

The UnpremultiplyColor overloads that return integral types will round
the value before it is cast to the appropriate type.