Support saving from the gray scale image mode
Gray scale images (with an optional alpha channel) will be encoded as a
monochrome AVIF.

The PremultiplyColor overload that returns a float will no longer round
the value to the nearest integer before returning it.
The rounding would cause incorrect behavior when processing floating
point image data in the [0, 1] range.

The PremultiplyColor overloads that return integral types will round
the value before it is cast to the appropriate type.