Add support for saving 32-bit HDR images
The format used for 32-bit images is the Rec. 2020 color space with the
Rec. 2100 PQ transfer function.
The Rec. 2100 HLG and SMPTE ST 428-1 transfer functions are not
currently supported.

ICC profiles are not supported for HDR images because the ICC profile
overrides the HDR color space and transfer functions.
Premultiplied alpha is supported, but its use will result in the image
data being clipped to the [0, 1] range.

Re-saving an AVIF HDR image will cause the saved image to be over
exposed, this is due to the PQ gamut adjustment that is made when
loading the image.
Without the gamut adjustment the loaded image will look too dark when
displayed in Photoshop.