Add a ColorTransferFunction::Clip value
This allows a 32-bit document to be saved without applying a HDR
curve.

In this mode the image data will be clipped to the [0, 1] range and
saved with a 1/13/6/full range NCLX profile (which are the same NCLX
values used for SDR images).
The NCLX values can also be overridden by an ICC profile, unlike the
Rec 2100 and SMPTE 428-1 transfer characteristics.