Convert the image to Rec. 2020 when saving as HDR
A best faith effort is made to detect if the image is already using a
Rec. 2020 color profile, the conversion will be skipped in that case.

If the image is not using a recognized Rec. 2020 color profile
Little CMS will be used to convert it to the Rec. 2020 color space.