Add a warning about HDR images and transparency
As far as I can tell, it is not very well defined how an application
should support HDR images with transparency.

While this plug-in can load and save HDR images that include
transparency, it may not be handled correctly by other applications.
It is also possible that this plug-in may not correctly handle HDR
images that include transparency from other applications.